+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Baseline Builder
 
 ## 1.7.0
+- Publish Baseline, swiftDialog and Installomator versions in new `Baseline-Builder` key of `BaselineConfig.plist`
+  - Configurable by `embed_versioning` parameter
 
 ## 1.6.0
 - Add support for creatign "faux" app icon for SimpleMDM